Refactor.  Start moving graphql requests into their own files/functions.
